project/requirements.txt
asyncpg==0.24.0
fastapi==0.68.1
sqlmodel==0.0.8
uvicorn==0.15.0
strawberry-graphql[fastapi]